data_scraper/requirements.txt
pymongo
requests
python-dotenv